+++ .travis_scripts/travis_install_dev.sh
@@ -1,4 +1,3 @@
 #!/bin/sh
 
-# docker exec app pip install black flake8 isort pytest
-docker exec app pip install pytest
+docker exec app pip install black flake8 isort pytest