+++ project/__init__.py
@@ -3,10 +3,12 @@ from flask import Flask
 from flask_cors import CORS
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
+from flask_bcrypt import Bcrypt
 
 # instantiate the dependencies
 db = SQLAlchemy()
 cors = CORS()
+bcrypt = Bcrypt()
 
 
 def create_app(script_info=None):
@@ -21,14 +23,15 @@ def create_app(script_info=None):
     # set up extensions
     db.init_app(app)
     cors.init_app(app, resources={r"*": {"origins": "*"}})
+    bcrypt.init_app(app)
 
     from project.api import api
 
     api.init_app(app)
 
     # migrate = Migrate(app, db)
-    Migrate(app, db)
     # Try also add migrate = Migrate() and migrate.init_app(app)
+    Migrate(app, db)
 
     # shell context for flask cli
     @app.shell_context_processor