+++ project/tests/unit/test_config.py
@@ -1,5 +1,5 @@
 # Test the configuration
-import os
+
 
 def test_testing_config(test_app):
     test_app.config.from_object("project.config.TestingConfig")