+++ CHANGELOG.md
@@ -1,2 +1,12 @@
-# 0.1.0
+# 0.2.0
+
+### Add
++ Add Flake8, Isort and Black
++ Add Postgres Database
++ Add tests for ping
++ Reenable all calls to DB
++ Add Bcrypt
++ Add `accounts` models and table
+
+## 0.1.0
 + First version